Exercise4

+++ src/main/java/at/fhj/msd/Calculator.java
@@ -2,11 +2,11 @@ package at.fhj.msd;
 
 public class Calculator {
 
-    public double add(double a, double b) {
+    public int add(int a, int b) {
         return a + b;
     }
 
-    public double minus(double a, double b) {
+    public int minus(int a, int b) {
         return a - b;
     }
 
@@ -14,7 +14,7 @@ public class Calculator {
         return a / b;
     }
 
-    public double multiply(double a, double b) {
+    public int multiply(int a, int b) {
         return a * b;
     }
 }
